backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary
gunicorn
python-dotenv
djangorestframework
whitenoise[brotli]